+++ composenav/src/main/java/com/adamkobus/compose/navigation/ui/NavComposableState.kt
@@ -1,0 +1,7 @@
+package com.adamkobus.compose.navigation.ui
+
+import androidx.compose.runtime.State
+
+internal data class NavComposableState(
+    val isInitialized: State<Boolean>
+)